backend/requirements-dev 2.txt
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
